requirements.txt
dbt-bigquery==1.8.4
google-cloud-bigquery==3.6.0
gspread==5.7.2
oauth2client==4.1.3
pandas==1.4.3
pathlib==1.0.1